+++ buildkite_waiter/src/cli.rs
@@ -1,6 +1,21 @@
 use url::Url;
 use structopt::StructOpt;
-use structopt::clap::{ArgGroup, AppSettings, arg_enum};
+use structopt::clap::AppSettings;
+
+fn allowed_build_states() -> &'static [&'static str] {
+    &[
+        "running",
+        "scheduled",
+        "passed",
+        "failed",
+        "blocked",
+        "canceled",
+        "canceling",
+        "skipped",
+        "not_run",
+        "finished",
+    ]
+}
 
 #[derive(StructOpt, Debug, PartialEq, Clone)]
 pub enum Command {
@@ -8,13 +23,33 @@ pub enum Command {
     Login,
     /// Remove the saved Buildkite API token
     Logout,
-    /// Wait for a build specified by URL
-    ByUrl,
+    /// Wait for a build specified by Buildkite web URL
+    ByUrl {
+        #[structopt(flatten)]
+        output: OutputArgs,
+        #[structopt(flatten)]
+        runtime: RuntimeArgs,
+        #[structopt(flatten)]
+        strategy: ByUrlStrategyArgs,
+    },
     /// Wait for a build specified by organization, pipeline and number
-    ByNumber,
+    ByNumber {
+        #[structopt(flatten)]
+        output: OutputArgs,
+        #[structopt(flatten)]
+        runtime: RuntimeArgs,
+        #[structopt(flatten)]
+        strategy: ByNumberStrategyArgs,
+    },
     /// Wait for the latest build matching certain filter criteria
-    Latest,
-    /// Wait for a build to finish
+    Latest {
+        #[structopt(flatten)]
+        output: OutputArgs,
+        #[structopt(flatten)]
+        runtime: RuntimeArgs,
+        #[structopt(flatten)]
+        strategy: LatestStrategyArgs,
+    },
     #[structopt(setting(AppSettings::Hidden), setting(AppSettings::TrailingVarArg), setting(AppSettings::AllowLeadingHyphen))]
     Wait {
         raw_parameters: Vec<String>
@@ -22,57 +57,51 @@ pub enum Command {
 }
 
 #[derive(StructOpt, Debug, PartialEq, Clone)]
-#[structopt(group = ArgGroup::with_name("build").required(true))]
-#[structopt(group = ArgGroup::with_name("output").required(false))]
-pub struct WaitArgs {
-    // Build selectors
-
-    #[structopt(long, group = "build")]
-    /// Specify build by the web URL
-    pub url: Option<Url>,
-    #[structopt(long, requires("pipeline"), group = "build")]
-    /// Specify the build number within the pipeline
-    pub number: Option<u32>,
-    #[structopt(long, group = "build")]
-    /// Find the latest build matching the other filters
-    pub latest: bool,
-
-    // Filters
-
-    #[structopt(long)]
-    /// Specify the organization containing the build
-    pub organization: Option<String>,
-    #[structopt(long, requires("organization"))]
-    /// Specify the pipeline containing the build
-    pub pipeline: Option<String>,
-    #[structopt(long, requires("latest"))]
-    /// Filter to builds on one or more branches
-    pub branch: Vec<String>,
-
-    // Behaviour
-
-    #[structopt(long, default_value = "480")]
-    /// Maximum time to wait for the build, in minutes
-    pub timeout: u32,
-    // TODO: nicer value format
-    #[structopt(long, possible_values = &ExitStatusStrategy::variants(), default_value = "CheckOnly", case_insensitive = true)]
-    /// Choose when to return an error exit code
-    pub error_on: ExitStatusStrategy,
-
-    // Outputs
-
-    #[structopt(long)]
-    /// Display a system notification when the app finishes waiting for a build
-    pub notification: bool,
-    #[structopt(long, group = "output")]
-    /// Output the contents of the default notification to stdout as a JSON object
-    pub output_notification_json: bool,
+pub struct ByNumberStrategyArgs {
+    pub organization: String,
+    pub pipeline: String,
+    pub number: u32,
 }
 
-arg_enum! {
-    #[derive(Debug, PartialEq, Clone, Copy)]
-    pub enum ExitStatusStrategy {
-        CheckOnly,
-        BuildFailedOrCanceled,
-    }
+#[derive(StructOpt, Debug, PartialEq, Clone)]
+pub struct ByUrlStrategyArgs {
+    pub url: Url,
+}
+
+#[derive(StructOpt, Debug, PartialEq, Clone)]
+pub struct LatestStrategyArgs {
+    #[structopt(long)]
+    pub organization: Option<String>,
+    #[structopt(long, requires("organization"))]
+    pub pipeline: Option<String>,
+    #[structopt(long)]
+    pub branch: Vec<String>,
+    #[structopt(long, conflicts_with("mine"))]
+    /// Find build by creator ID
+    pub creator: Option<String>,
+    #[structopt(long)]
+    /// Find build by (long) commit hash
+    pub commit: Option<String>,
+    #[structopt(long, possible_values = allowed_build_states())]
+    pub state: Vec<String>
+}
+
+#[derive(StructOpt, Debug, PartialEq, Clone)]
+pub struct RuntimeArgs {
+    #[structopt(long, default_value = "3600")]
+    /// Maximum time to wait for the build, in seconds
+    pub timeout: u32,
+
+    // TODO: expose or hardcode?
+    #[structopt(skip)]
+    pub request_cooldown: Option<u32>,
+}
+
+#[derive(StructOpt, Debug, PartialEq, Clone)]
+pub struct OutputArgs {
+    #[structopt(long)]
+    /// Send a system notification on completion
+    pub notification: bool,
+    #[structopt(long, possible_values = &["notification-json", "none"], default_value = "none")]
+    pub output: String,
 }