+++ buildkite_waiter/src/login.rs
@@ -3,7 +3,7 @@ use crate::api_auth;
 
 pub fn login() -> anyhow::Result<i32> {
     eprintln!("Generate an API Access Token at https://buildkite.com/user/api-access-tokens/new");
-    eprintln!(r#"Ensure you enable "Read Builds", and optionally "Read User"."#);
+    eprintln!(r#"Ensure you enable all relevant organizations, and enable the "Read Builds" (and optionally "Read User") permissions."#);
 
     let access_token = dialoguer::Password::new()
         .with_prompt("Buildkite API Access Token")